Moved file functions into separate file.
