+++ socketapi/CMakeLists.txt
@@ -6,6 +6,7 @@ LIST(APPEND neat_socketapi_headers
 
 LIST(APPEND neat_socketapi_sources
    neat-socketapi.c
+   neat-socketapi-file.c
    neat-socketapi-internals.c
    neat-socketapi-io.c
    neat-socketapi-poll.c
